Renamed script, created sensor functionality
